requirements.txt
beautifulsoup4==4.8.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
colorama==0.4.1
Flask==1.1.1
future==0.18.2
idna==2.5
itsdangerous==1.1.0
Jinja2==2.10.3
line-bot-sdk==1.15.0
lxml==4.4.2
MarkupSafe==1.1.1
prompt-toolkit==2.0.10
Pygments==2.4.2
requests==2.22.0
requests2==2.16.0
six==1.13.0
soupsieve==1.9.5
urllib3==1.21.1
wcwidth==0.1.7
Werkzeug==0.16.0
